Put material samplers in separate files
