+++ src/appleseed/renderer/kernel/lighting/directlightingintegrator.cpp
@@ -84,230 +84,6 @@ namespace renderer
 //   compute_incoming_radiance
 //
 
-BSDFSampler::BSDFSampler(
-    const BSDF&          bsdf,
-    const void*          bsdf_data,
-    const int            bsdf_sampling_modes,
-    const ShadingPoint&  shading_point)
-        : m_bsdf(bsdf)
-        , m_bsdf_data(bsdf_data)
-        , m_bsdf_sampling_modes(bsdf_sampling_modes)
-        , m_shading_basis(shading_point.get_shading_basis())
-        , m_geometric_normal(shading_point.get_geometric_normal())
-        , m_shading_point(shading_point)
-{
-}
-
-const Vector3d& BSDFSampler::get_point() const
-{
-    return m_shading_point.get_point();
-}
-
-bool BSDFSampler::contributes_to_light_sampling() const
-{
-    return !m_bsdf.is_purely_specular();
-}
-
-float BSDFSampler::trace_between(
-    const ShadingContext& shading_context,
-    const Vector3d& target_position
-    ) const
-{
-    return shading_context.get_tracer().trace_between(
-        m_shading_point,
-        target_position,
-        VisibilityFlags::ShadowRay
-        );
-}
-
-bool BSDFSampler::sample(
-    SamplingContext&   sampling_context,
-    const Dual3d&      outgoing,
-    Dual3f&            incoming,
-    Spectrum&          value,
-    float&             pdf) const
-{
-    // Sample the BSDF.
-    BSDFSample sample(&m_shading_point, Dual3f(outgoing));
-    m_bsdf.sample(
-        sampling_context,
-        m_bsdf_data,
-        false,                  // not adjoint
-        true,                   // multiply by |cos(incoming, normal)|
-        m_bsdf_sampling_modes,
-        sample);
-
-    // Filter scattering modes.
-    if (!(m_bsdf_sampling_modes & sample.m_mode))
-        return false;
-
-    incoming = sample.m_incoming;
-    value = sample.m_value;
-    pdf = sample.m_probability;
-
-    return true;
-}
-
-float BSDFSampler::evaluate(
-    int light_sampling_modes,
-    Vector3f outgoing,
-    Vector3f incoming,
-    Spectrum& value
-    ) const
-{
-    // Evaluate the BSDF.
-    return m_bsdf.evaluate(
-        m_bsdf_data,
-        false,              // not adjoint
-        true,               // multiply by |cos(incoming, normal)|
-        Vector3f(m_geometric_normal),
-        Basis3f(m_shading_basis),
-        Vector3f(outgoing),
-        Vector3f(incoming),
-        light_sampling_modes,
-        value);
-}
-
-const ShadingPoint& BSDFSampler::trace(
-    const ShadingContext&     shading_context,
-    const Vector3f&           direction,
-    float&                    transmission) const
-{
-    return shading_context.get_tracer().trace(
-        m_shading_point,
-        Vector3d(direction),
-        VisibilityFlags::ShadowRay,
-        transmission);
-}
-
-bool BSDFSampler::cull_incoming_direction(const Vector3d& incoming) const
-{
-    // Cull light samples behind the shading surface if the BSDF is either reflective or transmissive,
-    // but not both.
-    if (m_bsdf.get_type() != BSDF::AllBSDFTypes)
-    {
-        double cos_in = dot(incoming, m_shading_basis.get_normal());
-        if (m_bsdf.get_type() == BSDF::Transmissive)
-            cos_in = -cos_in;
-        if (cos_in <= 0.0)
-            return true;
-    }
-    return false;
-}
-
-PhaseFunctionSampler::PhaseFunctionSampler(
-    const ShadingRay&        volume_ray,
-    const PhaseFunction&     phasefunction,
-    const void*              phasefunction_data,
-    const float              distance)
-      : m_volume_ray(volume_ray)
-      , m_phasefunction(phasefunction)
-      , m_phasefunction_data(phasefunction_data)
-      , m_distance(distance)
-      , m_point(m_volume_ray.point_at(distance))
-{
-}
-
-const Vector3d& PhaseFunctionSampler::get_point() const
-{
-    return m_point;
-}
-
-bool PhaseFunctionSampler::contributes_to_light_sampling() const
-    {
-        return true;
-    }
-
-bool PhaseFunctionSampler::sample(
-    SamplingContext&   sampling_context,
-    const Dual3d&      outgoing,
-    Dual3f&            incoming,
-    Spectrum&          value,
-    float&             pdf) const
-{
-    Vector3f incoming_direction;
-
-    // Sample the Phase Function.
-    pdf = m_phasefunction.sample(
-        sampling_context,
-        m_volume_ray,
-        m_phasefunction_data,
-        m_distance,
-        incoming_direction
-        );
-
-    incoming = Dual3f(incoming_direction);
-
-    m_phasefunction.scattering_coefficient(
-        m_volume_ray, m_phasefunction_data, m_distance, value);
-
-    Spectrum transmission;
-    m_phasefunction.evaluate_transmission(
-        m_volume_ray, m_phasefunction_data, m_distance, transmission);
-    value *= transmission;
-
-    return true;
-}
-
-float PhaseFunctionSampler::evaluate(
-    int light_sampling_modes,
-    Vector3f outgoing,
-    Vector3f incoming,
-    Spectrum& value
-    ) const
-{
-    // Evaluate the Phase Function.
-    float pdf = m_phasefunction.evaluate(
-        m_volume_ray,
-        m_phasefunction_data,
-        m_distance,
-        incoming);
-
-    m_phasefunction.scattering_coefficient(
-        m_volume_ray, m_phasefunction_data, m_distance, value);
-
-    Spectrum transmission;
-    m_phasefunction.evaluate_transmission(
-        m_volume_ray, m_phasefunction_data, m_distance, transmission);
-    value *= transmission;
-
-    return pdf;
-}
-
-float PhaseFunctionSampler::trace_between(
-    const ShadingContext& shading_context,
-    const Vector3d& target_position
-    ) const
-{
-    return shading_context.get_tracer().trace_between(
-        m_point,
-        target_position,
-        m_volume_ray.m_time,
-        VisibilityFlags::ShadowRay,
-        m_volume_ray.m_depth + 1
-        );
-}
-
-const ShadingPoint& PhaseFunctionSampler::trace(
-    const ShadingContext&     shading_context,
-    const Vector3f&           direction,
-    float&                    transmission) const
-{
-    return shading_context.get_tracer().trace(
-        m_point,
-        Vector3d(direction),
-        m_volume_ray.m_time,
-        VisibilityFlags::ShadowRay,
-        m_volume_ray.m_depth + 1,
-        transmission);
-}
-
-bool PhaseFunctionSampler::cull_incoming_direction(const Vector3d& incoming) const
-{
-    // No culling for phase function
-    return false;
-}
-
 DirectLightingIntegrator::DirectLightingIntegrator(
     const ShadingContext&       shading_context,
     const LightSampler&         light_sampler,
