+++ src/appleseed/renderer/kernel/lighting/materialsamplers.h
@@ -1,0 +1,191 @@
+
+//
+// This source file is part of appleseed.
+// Visit http://appleseedhq.net/ for additional information and resources.
+//
+// This software is released under the MIT license.
+//
+// Copyright (c) 2017 Artem Bishev, The appleseedhq Organization
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+
+#ifndef APPLESEED_RENDERER_KERNEL_LIGHTING_MATERIALSAMPLERS_H
+#define APPLESEED_RENDERER_KERNEL_LIGHTING_MATERIALSAMPLERS_H
+
+// appleseed.renderer headers.
+#include "renderer/global/globaltypes.h"
+#include "renderer/kernel/shading/shadingray.h"
+
+// appleseed.foundation headers.
+#include "foundation/math/basis.h"
+#include "foundation/math/dual.h"
+#include "foundation/math/vector.h"
+
+// Standard headers.
+#include <cstddef>
+
+// Forward declarations.
+namespace renderer  { class BSDF; }
+namespace renderer  { class ShadingContext; }
+namespace renderer  { class ShadingPoint; }
+
+//
+// This file contains wrappers over samplable functions, such as BSDF and PhaseFunction,
+// These wrappers provide a uniform interface for sampling and evaluating these functions
+// in a certain point, which can be used during lighting integration.
+//
+
+namespace renderer
+{
+
+class IMaterialSampler
+{
+  public:
+    virtual const foundation::Vector3d& get_point() const = 0;
+
+    virtual bool contributes_to_light_sampling() const = 0;
+
+    virtual const ShadingPoint& trace(
+        const ShadingContext&        shading_context,
+        const foundation::Vector3f&  direction,
+        float&                       transmission) const = 0;
+
+    virtual float trace_between(
+        const ShadingContext& shading_context,
+        const foundation::Vector3d& target_position
+        ) const = 0;
+
+    virtual bool sample(
+        SamplingContext&   sampling_context,
+        const foundation::Dual3d&      outgoing,
+        foundation::Dual3f&            incoming,
+        Spectrum&          value,
+        float&             pdf) const = 0;
+
+    virtual float evaluate(
+        int                light_sampling_modes,
+        foundation::Vector3f           outgoing,
+        foundation::Vector3f           incoming,
+        Spectrum&          value
+        ) const = 0;
+
+    virtual bool cull_incoming_direction(
+        const foundation::Vector3d& incoming) const = 0;
+};
+
+class BSDFSampler : public IMaterialSampler
+{
+  public:
+    BSDFSampler(
+        const BSDF&          bsdf,
+        const void*          bsdf_data,
+        const int            bsdf_sampling_modes,
+        const ShadingPoint&  shading_point);
+
+    virtual const foundation::Vector3d& get_point() const;
+
+    virtual bool contributes_to_light_sampling() const;
+
+    virtual float trace_between(
+        const ShadingContext&         shading_context,
+        const foundation::Vector3d&   target_position
+        ) const;
+
+    virtual bool sample(
+        SamplingContext&             sampling_context,
+        const foundation::Dual3d&    outgoing,
+        foundation::Dual3f&          incoming,
+        Spectrum&                    value,
+        float&                       pdf) const;
+
+    virtual float evaluate(
+        int                     light_sampling_modes,
+        foundation::Vector3f    outgoing,
+        foundation::Vector3f    incoming,
+        Spectrum&               value
+        ) const;
+
+    virtual const ShadingPoint& trace(
+        const ShadingContext&          shading_context,
+        const foundation::Vector3f&    direction,
+        float&                         transmission) const override;
+
+    virtual bool cull_incoming_direction(const foundation::Vector3d& incoming) const;
+
+private:
+    const BSDF&                      m_bsdf;
+    const void*                      m_bsdf_data;
+    const int                        m_bsdf_sampling_modes;
+    const foundation::Basis3d&       m_shading_basis;
+    const foundation::Vector3d&      m_geometric_normal;
+    const ShadingPoint&              m_shading_point;
+};
+
+class PhaseFunctionSampler : public IMaterialSampler
+{
+  public:
+    PhaseFunctionSampler(
+        const ShadingRay&        volume_ray,
+        const PhaseFunction&     phasefunction,
+        const void*              phasefunction_data,
+        const float              distance);
+
+    virtual const foundation::Vector3d& get_point() const;
+
+    virtual bool contributes_to_light_sampling() const;
+
+    virtual bool sample(
+        SamplingContext&           sampling_context,
+        const foundation::Dual3d&  outgoing,
+        foundation::Dual3f&        incoming,
+        Spectrum&                  value,
+        float&                     pdf) const;
+
+    virtual float evaluate(
+        int                   light_sampling_modes,
+        foundation::Vector3f  outgoing,
+        foundation::Vector3f  incoming,
+        Spectrum&             value
+        ) const;
+
+    virtual float trace_between(
+        const ShadingContext&        shading_context,
+        const foundation::Vector3d&  target_position
+        ) const;
+
+    virtual const ShadingPoint& trace(
+        const ShadingContext&        shading_context,
+        const foundation::Vector3f&  direction,
+        float&                       transmission) const override;
+
+    virtual bool cull_incoming_direction(
+        const foundation::Vector3d&  incoming) const;
+
+private:
+    const ShadingRay&          m_volume_ray;
+    const PhaseFunction&       m_phasefunction;
+    const void*                m_phasefunction_data;
+    const float                m_distance;
+    const foundation::Vector3d m_point;
+};
+
+}       // namespace renderer
+
+#endif  // !APPLESEED_RENDERER_KERNEL_LIGHTING_MATERIALSAMPLERS_H
